% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurnThin_function.R
\name{BurnThin}
\alias{BurnThin}
\title{Burning and thinning the MCMC samples.}
\usage{
BurnThin(bac, burn = 0, thin = 1)
}
\arguments{
\item{bac}{The fit from the BAC function.}

\item{burn}{How many of the initial samples should be dropped. Default is 0.}

\item{thin}{One every how many samples should be kept. Defaults to 1.}
}
\description{
Burning and thinning the MCMC samples.
}
