% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcLogLike_function.R
\name{CalcLogLike}
\alias{CalcLogLike}
\title{Marginal log likelihood of linear model.
Function that calculates the marginal log likelihood of a normal outcome on a design
matrix with normal prior on the coefficients and inverse gamma prior on the
variance.}
\usage{
CalcLogLike(outcome, design_mat, nu_prior, lambda_prior, mu_prior, Sigma_prior)
}
\arguments{
\item{outcome}{Vector with the entries of the outcome.}

\item{design_mat}{The design matrix of the outcome model.}

\item{nu_prior}{The value of nu in the inverse gamma prior of sigma square.}

\item{lambda_prior}{The value of lambda in the inverse gamma prior of sigma square.}

\item{mu_prior}{The mean of the normal prior on the coefficients of the model.}

\item{Sigma_prior}{The covariance matrix of the normal prior on the coefficients of
the outcome model.}
}
\description{
Marginal log likelihood of linear model.
Function that calculates the marginal log likelihood of a normal outcome on a design
matrix with normal prior on the coefficients and inverse gamma prior on the
variance.
}

