\name{ER}
\docType{data}
\alias{ER}
\title{Chromosome-Wide Mapping of Estrogen Receptor Binding Reveals Long-Range Regulation Requiring the Forkhead Protein FoxA1}
\description{
  This is a subset of the data containing 30000 probes on chromosome 21.
}
\usage{data(ER)}
\source{http://www.cell.com/content/article/abstract?uid=PIIS0092867405004538}
\references{Cell, Vol 122, 33-43, 15 July 2005}
\keyword{datasets}
