% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPriorOdds_function.R
\name{LogPriorOdds}
\alias{LogPriorOdds}
\title{Log Prior Odds for the BAC inclusions indicators.
Function that calculates the prior odds for the MH move through the model space.}
\usage{
LogPriorOdds(proposed, current, other_model, out_model, omega)
}
\arguments{
\item{proposed}{Vector of 0's and 1's corresponding to proposed covariates that will
be included in the model. 0 corresponds to exclusion and 1 to inclusion of the
the covariate accordingly.}

\item{current}{Same length with proposed. Includes the current values of the
inclusion indicators.}

\item{other_model}{Same length as proposed and current. Includes the current values
of the inclusion indicators for the other model. For example, when we are updating
the outcome model, other_model includes the inclusion indicators for the exposure.}

\item{out_model}{Logical. Set to TRUE if we are updating the outcome model.}

\item{omega}{The omega of the BAC prior.}
}
\description{
Log Prior Odds for the BAC inclusions indicators.
Function that calculates the prior odds for the MH move through the model space.
}

