% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeArrays_function.R
\name{MakeArrays}
\alias{MakeArrays}
\title{Arrays to save MCMC posterior samples.}
\usage{
MakeArrays(
  chains,
  Nsims,
  num_conf,
  starting_alphas,
  starting_coefs,
  starting_vars,
  omega = 50000,
  mu_priorX,
  mu_priorY,
  Sigma_priorX,
  Sigma_priorY,
  alpha_priorX,
  alpha_priorY,
  beta_priorX,
  beta_priorY
)
}
\arguments{
\item{chains}{The number of MCMC chains.}

\item{Nsims}{The number of posterior samples per chain.}

\item{num_conf}{The number of potential confounders.}

\item{starting_alphas}{Array of dimensions: model (exposure or outcome),
chains, potential confounders. Entries 0/1 represent exclusion/inclusion of
a covariate in the model. If NULL, values are set from the prior.}

\item{starting_coefs}{Array with the starting values of all coefficients.
Dimensions are: Exposure/Outcome model, chains, and covariate (intercept,
coefficient of exposure, covariates). The coefficient of exposure should be
NA for the exposure model. If NULL, values are set from the normal with mean
zero and variance that of the prior divided by 50 ^ 2.}

\item{starting_vars}{Array including the starting values for the residual
variances. Dimensions correspond to: Exposure/Outcome model, and chains. If
NULL, values are set from an inverse gamma with parameters alpha and beta
set to the prior values times 200.}

\item{omega}{The omega of the BAC prior on inclusion indicators.}

\item{mu_priorX}{The mean of the normal prior on the coefficients of the
exposure model, where the first element corresponds to the intercept, and
the remaining to the coefficients of the columns in D.}

\item{mu_priorY}{The mean of the normal prior on the coefficients of the
outcome model, where the first element corresponds to the intercept, the
second to the exposure, and the remaining to the coefficients in from of the
columns in D.}

\item{Sigma_priorX}{The covariance matrix of the normal prior on the
coefficients of the exposure model.}

\item{Sigma_priorY}{The covariance matrix of the normal prior on the
coefficients of the outcome model.}

\item{alpha_priorX}{The value of alpha in the inverse gamma prior for the
residual variance of the exposure model.}

\item{alpha_priorY}{The value of alpha in the inverse gamma prior for the
residual variance of the outcome model.}

\item{beta_priorX}{The value of beta in the inverse gamma prior for the
residual variance of the exposure model.}

\item{beta_priorY}{The value of beta in the inverse gamma prior for the
residual variance of the outcome model.}
}
\description{
Arrays to save MCMC posterior samples.
}
