% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelWeights_function.R
\name{ModelWeights}
\alias{ModelWeights}
\title{Posterior weights of all models in MCMC.}
\usage{
ModelWeights(bac, model = c("Outcome", "Exposure"))
}
\arguments{
\item{bac}{The BAC fit.}

\item{model}{Options are 'Outcome' or 'Exposure'. Defines which model we 
want to look at.}
}
\description{
Function that calculates and orders the models that were chosen in the MCMC
at least once.
}
