% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostQuants_function.R
\name{PostQuants}
\alias{PostQuants}
\title{Calculates the posterior mean and standard deviation of a coefficient.}
\usage{
PostQuants(predictor, resid, prior_mean, prior_sd, resid_var)
}
\arguments{
\item{predictor}{Vector including the values of the predictor for which the
posterior quantities are calculated.}

\item{resid}{The residuals of the outcome on covariates excluding intercept.}

\item{prior_mean}{The prior mean for the intercept.}

\item{prior_sd}{The prior standard deviation of the intercept.}

\item{resid_var}{The variance of the residuals.}
}
\description{
Calculates the posterior mean and standard deviation of a coefficient.
}
