% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{Toy data set.}
\format{A matrix with 100 rows and 52 columns.
\describe{
  \item{Y}{Continuous outcome.}
  \item{X}{Continuous exposure}
  \item{D1, D2, ..., D50}{Possible confounders.}
}}
\usage{
toyData
}
\description{
A data set containing simulated data with a continuous exposure, outcome and
covariates.
}
\keyword{datasets}
