% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateAlphaX_function.R
\name{UpdateAlphaX}
\alias{UpdateAlphaX}
\title{Updating the outcome inclusion indicators
Function that updates the indicators of inclusion for predictors in the outcome
model.}
\usage{
UpdateAlphaX(X, D, curr_alphaX, curr_alphaY, nu_prior, lambda_prior, mu_prior,
  Sigma_prior, omega)
}
\arguments{
\item{X}{Vector of the treatment}

\item{D}{Matrix or data frame where the columns correspond to all possible
predictors of Y, and the rows correspond to the units.}

\item{curr_alphaX}{Vector of 0's and 1's of length equal to the number of columns
in D. Variables in D with corresponding curr_alphaX entry equal to 1 are in the
exposure model at the specific iteration of BAC. The proposed value of alphaX will
be the same as curr_alphaX by changing only one 0 to 1 or one 1 to 0.}

\item{curr_alphaY}{Vector of 0's and 1's of length equal to the number of columns
in D. Only variables in D with corresponding curr_alphaY entry equal to 1 are 
in the outcome model at the specific iteration of BAC.}

\item{nu_prior}{The value of nu in the prior of sigma square for the exposure model.}

\item{lambda_prior}{The value of lambda in the prior of sigma square for the
exposure model.}

\item{mu_prior}{The mean of the normal prior on the coefficients of the exposure
model, where the first element corresponds to the intercept, and the remaining to
the coefficients in from of the columns in D. The length of this vector must be
equal to 1 + the number of columns in D.}

\item{Sigma_prior}{The covariance matrix of the normal prior on the coefficients of
the exposure model. The structure should be in the same order as mu_prior and the
matrix must be symmetric with dimensions equal to 1 + the number of columns in D.}

\item{omega}{The omega parameter of the BAC prior.}
}
\description{
Updating the outcome inclusion indicators
Function that updates the indicators of inclusion for predictors in the outcome
model.
}

