% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateVariances_function.R
\name{UpdateVariances}
\alias{UpdateVariances}
\title{MCMC update of the residual variance.}
\usage{
UpdateVariances(X, Y, D, current_alphas, current_coefs, alpha_priorX,
  alpha_priorY, beta_priorX, beta_priorY)
}
\arguments{
\item{X}{Vector of the treatment}

\item{Y}{Vector of the outcome.}

\item{D}{Matrix or data frame where the columns correspond to all possible
predictors of Y, and the rows correspond to the units.}

\item{current_alphas}{Matrix of inclusion indicators with rows corresponding
to exposure and outcome models and columns to covariates.}

\item{current_coefs}{Matrix of coefficients with rows corresponding to
exposure and outcome models and columns to intercept, exposure and
covariates.}

\item{alpha_priorX}{The value of alpha in the inverse gamma prior for the
residual variance of the exposure model.}

\item{alpha_priorY}{The value of alpha in the inverse gamma prior for the
residual variance of the outcome model.}

\item{beta_priorX}{The value of beta in the inverse gamma prior for the
residual variance of the exposure model.}

\item{beta_priorY}{The value of beta in the inverse gamma prior for the
residual variance of the outcome model.}
}
\description{
MCMC update of the residual variance.
}
