% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAC_function.R
\name{BAC}
\alias{BAC}
\title{Bayesian Adjustment for Confounding
Function that fits the BAC method of Wang, Parmigianni, Dominici (2012).}
\usage{
BAC(X, Y, D, Nsims, mu_priorX = NULL, mu_priorY = NULL, nu_priorX = 0.01,
  nu_priorY = 0.01, lambda_priorX = 0.01, lambda_priorY = 0.01,
  Sigma_priorX = NULL, Sigma_priorY = NULL, omega = 50000)
}
\arguments{
\item{X}{Vector of the treatment}

\item{Y}{Vector of the outcome.}

\item{D}{Matrix or data frame where the columns correspond to all possible
predictors of Y, and the rows correspond to the units.}

\item{Nsims}{Number of posterior samples we wish to get.}

\item{mu_priorX}{The mean of the normal prior on the coefficients of the exposure
model, where the first element corresponds to the intercept, and the remaining to
the coefficients in from of the columns in D. The length of this vector must be
equal to 1 + the number of columns in D.}

\item{mu_priorY}{The mean of the normal prior on the coefficients of the outcome
model, where the first element corresponds to the intercept, the second to the
treatment, and the remaining to the coefficients in from of the columns in D. The
length of this vector must be equal to 2 + the number of columns in D.}

\item{nu_priorX}{The value of nu in the prior of sigma square for the exposure
model. Defaults to 0.01.}

\item{nu_priorY}{The value of nu in the prior of sigma square for the outcome model.
Defaults to 0.01.}

\item{lambda_priorX}{The value of lambda in the prior of sigma square for the
exposure model. Defaults to 0.01.}

\item{lambda_priorY}{The value of lambda in the prior of sigma square for the
outcome model. Defaults to 0.01.}

\item{Sigma_priorX}{The covariance matrix of the normal prior on the coefficients of
the exposure model. The structure should be in the same order as mu_priorX and the
matrix must be symmetric with dimensions equal to 1 + the number of columns in D.}

\item{Sigma_priorY}{The covariance matrix of the normal prior on the coefficients
of the outcome model. The structure should be in the same order as mu_priorY and the
matrix must be symmetric with dimensions equal to 2 + the number of columns in D.}

\item{omega}{The omega parameter of the BAC prior. Defaults to 50000.}
}
\description{
Bayesian Adjustment for Confounding
Function that fits the BAC method of Wang, Parmigianni, Dominici (2012).
}

